build: move parameter tests to api to try fix the tests issue
