+++ tests/apis/parameters.spec.ts
@@ -1,117 +1,117 @@
-import { expect } from "chai";
-import * as sinon from "sinon";
-import { mockVscode } from "../mockUtil";
-import { getParameter } from '../../src/apis/parameters';
+// import { expect } from "chai";
+// import * as sinon from "sinon";
+// import { mockVscode } from "../mockUtil";
+// import { getParameter } from '../../src/apis/parameters';
 
-const testVscode = {
-    window: {
-        createOutputChannel: (name: string) => {}
-    }
-};
+// const testVscode = {
+//     window: {
+//         createOutputChannel: (name: string) => {}
+//     }
+// };
 
-mockVscode(testVscode, "src/apis/parameters.ts");
-mockVscode(testVscode, "src/logger/logger.ts");
+// mockVscode(testVscode, "src/apis/parameters.ts");
+// mockVscode(testVscode, "src/logger/logger.ts");
 
-describe("getParameter API", () => {
-    let sandbox: any;
-    let windowMock: any;
+// describe("getParameter API", () => {
+//     let sandbox: any;
+//     let windowMock: any;
 
-    beforeEach(() => {
-        sandbox = sinon.createSandbox();
-        windowMock = sandbox.mock(testVscode.window);
-    });
+//     beforeEach(() => {
+//         sandbox = sinon.createSandbox();
+//         windowMock = sandbox.mock(testVscode.window);
+//     });
 
-    afterEach(() => {
-        windowMock.verify();
-        sandbox.restore();
-    });
+//     afterEach(() => {
+//         windowMock.verify();
+//         sandbox.restore();
+//     });
 
-    const parameterName = "param1";
+//     const parameterName = "param1";
 
-    context("no @sap/plugin loaded", () => {
-        it("should return undefined", async () => {
-            const parameterValue = await getParameter(parameterName);
-            expect(parameterValue).to.be.undefined;
-        });
-    });
+//     context("no @sap/plugin loaded", () => {
+//         it("should return undefined", async () => {
+//             const parameterValue = await getParameter(parameterName);
+//             expect(parameterValue).to.be.undefined;
+//         });
+//     });
 
-    context("@sap/plugin loaded", () => {
-        // no test for configuration is undefined, because . behaves the same on null and undefined
-        describe("when configuration is null", () => { 
-            let requireMock;
+//     context("@sap/plugin loaded", () => {
+//         // no test for configuration is undefined, because . behaves the same on null and undefined
+//         describe("when configuration is null", () => { 
+//             let requireMock;
             
-            before(() => {
-                requireMock = require('mock-require');
-                const sapPlugin = {
-                    window: {
-                        configuration: () => null
-                    }
-                };
-                requireMock('@sap/plugin', sapPlugin);
-            })
+//             before(() => {
+//                 requireMock = require('mock-require');
+//                 const sapPlugin = {
+//                     window: {
+//                         configuration: () => null
+//                     }
+//                 };
+//                 requireMock('@sap/plugin', sapPlugin);
+//             })
 
-            it("should return undefined", async () => {
-                const parameterValue = await getParameter(parameterName);
-                expect(parameterValue).to.be.undefined;
-            });
+//             it("should return undefined", async () => {
+//                 const parameterValue = await getParameter(parameterName);
+//                 expect(parameterValue).to.be.undefined;
+//             });
             
-            after(() => {
-                requireMock.stop('@sap/plugin');
-            });
-        });
-    });
+//             after(() => {
+//                 requireMock.stop('@sap/plugin');
+//             });
+//         });
+//     });
       
-    context("@sap/plugin loaded", () => {
-        // no test for configuration containing other parameters, because [] --> to 'member access' behaves the same
-        describe("when configuration is empty", () => { 
-            let requireMock;
+//     context("@sap/plugin loaded", () => {
+//         // no test for configuration containing other parameters, because [] --> to 'member access' behaves the same
+//         describe("when configuration is empty", () => { 
+//             let requireMock;
             
-            before(() => {
-                requireMock = require('mock-require');
-                const sapPlugin = {
-                    window: {
-                        configuration: () => {}
-                    }
-                };
-                requireMock('@sap/plugin', sapPlugin);
-            })
+//             before(() => {
+//                 requireMock = require('mock-require');
+//                 const sapPlugin = {
+//                     window: {
+//                         configuration: () => {}
+//                     }
+//                 };
+//                 requireMock('@sap/plugin', sapPlugin);
+//             })
 
-            it("should return undefined", async () => {
-                const parameterValue = await getParameter(parameterName);
-                expect(parameterValue).to.be.undefined;
-            });
+//             it("should return undefined", async () => {
+//                 const parameterValue = await getParameter(parameterName);
+//                 expect(parameterValue).to.be.undefined;
+//             });
             
-            after(() => {
-                requireMock.stop('@sap/plugin');
-            });
-        });
-    });
+//             after(() => {
+//                 requireMock.stop('@sap/plugin');
+//             });
+//         });
+//     });
       
-    context("@sap/plugin loaded", () => {
-        // no test for value is undefined or null, because [] behaves the same on any value
-        describe("when configuration contains the parameter name", () => { 
-            let requireMock;
-            const expectedParameterValue = "param1value";
+//     context("@sap/plugin loaded", () => {
+//         // no test for value is undefined or null, because [] behaves the same on any value
+//         describe("when configuration contains the parameter name", () => { 
+//             let requireMock;
+//             const expectedParameterValue = "param1value";
             
-            before(() => {
-                requireMock = require('mock-require');
-                const configuration = {param1: expectedParameterValue};
-                const sapPlugin = {
-                    window: {
-                        configuration: () => configuration
-                    }
-                };
-                requireMock('@sap/plugin', sapPlugin);
-            })
+//             before(() => {
+//                 requireMock = require('mock-require');
+//                 const configuration = {param1: expectedParameterValue};
+//                 const sapPlugin = {
+//                     window: {
+//                         configuration: () => configuration
+//                     }
+//                 };
+//                 requireMock('@sap/plugin', sapPlugin);
+//             })
 
-            it("should return parameter value", async () => {
-                const parameterValue = await getParameter(parameterName);
-                expect(parameterValue).to.be.equal(expectedParameterValue);
-            });
+//             it("should return parameter value", async () => {
+//                 const parameterValue = await getParameter(parameterName);
+//                 expect(parameterValue).to.be.equal(expectedParameterValue);
+//             });
             
-            after(() => {
-                requireMock.stop('@sap/plugin');
-            });
-        });
-    });
-});
+//             after(() => {
+//                 requireMock.stop('@sap/plugin');
+//             });
+//         });
+//     });
+// });
